+++ src/lib/url-params.js
@@ -3,7 +3,19 @@
 var querystring = require("./querystring");
 
 function getUrlParams() {
-  return querystring.parse(window.location.href);
+  var targetWindow = window;
+
+  // If in same-origin iframe, prefer top window (where redirect params are)
+  try {
+    if (window.top && window.top !== window && window.top.location) {
+      targetWindow = window.top;
+    }
+    // eslint-disable-next-line no-unused-vars
+  } catch (_e) {
+    // Cross-origin iframe - fall back to current window
+  }
+
+  return querystring.parse(targetWindow.location.href);
 }
 
 module.exports = {